+++ ECU_Manager/Packets/PK_CriticalMemoryAcknowledge.cs
@@ -28,7 +28,7 @@ namespace ECU_Manager.Packets
             ErrorCode = (uint)errorcode;
 
             PacketLength = 0;
-            PacketLength = (byte)Marshal.SizeOf(GetType());
+            PacketLength = (ushort)Marshal.SizeOf(GetType());
         }
 
     }
