+++ ECU_Manager/Packets/PK_DragPointResponse.cs
@@ -32,7 +32,7 @@ namespace ECU_Manager.Packets
             Point = new DragPoint();
 
             PacketLength = 0;
-            PacketLength = (byte)Marshal.SizeOf(GetType());
+            PacketLength = (ushort)Marshal.SizeOf(GetType());
         }
 
     }
